requirements.txt
pysqlite==2.8.3
chalice==1.0.2
